Move countdown business code in good directory
